[refactor] moved out some test stuff to different file
